// src/app/shared/interfaces/user.interface.ts
// Generated by https://quicktype.io
export interface Username {
    id: string;
    user: string;
    created_at: string;
    updated_at: string;
    deleted_at: null;
    state: number;
    charge_id: string;
    person_id: string;
    token: string;
}
// Generated by https://quicktype.io

export interface Person {
    id: string;
    number_person: string;
    name: string;
    father_last_name: string;
    mother_last_name: string;
    full_name: string;
    ci: string;
    birth_date: string;
    birth_place: string;
    phone: string;
    cell_phone: string;
    address: string;
    email: string;
    latitude: string;
    longitude: string;
    created_at: string;
    updated_at: string;
    deleted_at: null;
    state: number;
    type_person_id: string;
    branch_office_id: string;
}